+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Fixed
+- Use props `children` in slots if slot children is empty.
+
 ## [8.126.10] - 2021-01-26
 
 ### Fixed